backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-snap7==2.0.0
pymcprotocol==0.3.0
pylogix==1.1.5
pymodbus==3.7.0
pytest==8.3.0
httpx==0.27.0
tomli==2.0.1; python_version < '3.11'
python-dotenv==1.0.1
